update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@62345 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.4/fs/ceph/file.c
@@ -1303,7 +1303,7 @@ static int ceph_zero_objects(struct inode *inode, loff_t offset, loff_t length)
 	s32 stripe_unit = ceph_file_layout_su(ci->i_layout);
 	s32 stripe_count = ceph_file_layout_stripe_count(ci->i_layout);
 	s32 object_size = ceph_file_layout_object_size(ci->i_layout);
-	u64 object_set_size = object_size * stripe_count;
+	u64 object_set_size = (u64) object_size * stripe_count;
 	u64 nearly, t;
 
 	/* round offset up to next period boundary */
